+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Shared from './pages/Shared';
-import FolderPage from './pages/FolderPage';
+import SharedPage from './pages/SharedPage/SharedPage';
+import FolderPage from './pages/FolderPage/FolderPage';
 import Layout from './pages/Layout';
 
 function App() {
@@ -9,7 +9,7 @@ function App() {
       <BrowserRouter>
         <Layout>
           <Routes>
-            <Route path="/shared" element={<Shared />} />
+            <Route path="/shared" element={<SharedPage />} />
             <Route path="/folder" element={<FolderPage />} />
             <Route path="*" element={<div>잘못된 페이지입니다.</div>} />
           </Routes>
