--- src/pages/FolderPage.jsx
@@ -1,11 +1,0 @@
-import FolderMain from '../components/FolderMain/FolderMain';
-
-function FolderPage() {
-  return (
-    <>
-      <FolderMain />
-    </>
-  );
-}
-
-export default FolderPage;
