--- src/pages/Shared.jsx
@@ -1,10 +1,0 @@
-import SharedMain from '../components/SharedMain/SharedMain';
-import Layout from './Layout';
-
-export default function Shared() {
-  return (
-    <>
-      <SharedMain />
-    </>
-  );
-}
