+++ src/pages/SharedPage/SharedPage.css
